APIの通信部分を共有する準備
