+++ ios/Parjob/Parjob/API/APIManager.swift
@@ -1,0 +1,9 @@
+//
+//  APIManager.swift
+//  Parjob
+//
+//  Created by 岩見建汰 on 2018/06/25.
+//  Copyright © 2018年 Kenta Iwami. All rights reserved.
+//
+
+import Foundation
